Reorganized project structure in order to make functions testable
